Refactor To User Use Query
