+++ src/pages/Dashboard/DashboardLayout.tsx
@@ -1,7 +1,8 @@
 import Icon from "../../components/common/Icon";
 import Navbar from "../../components/common/Navbar/Navbar";
+import { UserContext } from "../../contexts/UserContext";
 import styles from "./DashboardLayout.module.scss";
-import { ReactNode } from "react";
+import { ReactNode, useContext } from "react";
 
 const dashboardLinks = [
 	{ href: "/dashboard", name: "Links", icon: <Icon variant='link' /> },
@@ -13,11 +14,13 @@ const dashboardLinks = [
 ];
 
 const DashboardLayout = ({ children }: { children: ReactNode }) => {
+	const { isUserPending } = useContext(UserContext);
+
 	return (
 		<div className={styles.dashboard_layout}>
 			<div>
 				<Navbar navigationLinks={dashboardLinks} className={styles.navbar} />
-				<main>{children}</main>
+				<main>{isUserPending ? <h1>Loading...</h1> : children}</main>
 			</div>
 		</div>
 	);
